+++ packages/fe-fpm-writer/src/page/common.ts
@@ -1,5 +1,8 @@
 import type { Editor } from 'mem-fs-editor';
+import { create as createStorage } from 'mem-fs';
+import { create } from 'mem-fs-editor';
 import { join } from 'path';
+import { render } from 'ejs';
 import type { ManifestNamespace } from '@sap-ux/project-access';
 import { validateBasePath } from '../common/validate';
 import type {
@@ -8,11 +11,19 @@ import type {
     InternalCustomPage,
     InternalObjectPage,
     ObjectPage,
+    ListReport,
     Navigation,
     InternalListReport
 } from './types';
 import type { Manifest } from '../common/types';
 import { FCL_ROUTER } from '../common/defaults';
+import { extendJSON } from '../common/file';
+import { getTemplatePath } from '../templates';
+
+type EnhancePageConfigFunction = (
+    data: ObjectPage | ListReport,
+    manifest: Manifest
+) => InternalObjectPage | InternalListReport;
 
 /**
  * Suffix for patterns to support arbitrary paramters
@@ -169,6 +180,45 @@ export function validatePageConfig(basePath: string, config: CustomPage | Object
             throw new Error(`Invalid routing configuration for navigation source ${config.navigation.sourcePage}!`);
         }
     }
+
+    return fs;
+}
+
+/**
+ * Add an generic page to an existing UI5 application.
+ * Supported pages - ListReport or ObjectPage.
+ *
+ * @param basePath - the base path
+ * @param data - the page configuration
+ * @param enhanceDataFn - Callback function for data enhancement
+ * @param templatePath - path to 'manifest.json' template
+ * @param fs - the memfs editor instance
+ * @returns the updated memfs editor instance
+ */
+export function extendPageJSON(
+    basePath: string,
+    data: ObjectPage,
+    enhanceDataFn: EnhancePageConfigFunction,
+    templatePath: string,
+    fs?: Editor
+): Editor {
+    if (!fs) {
+        fs = create(createStorage());
+    }
+    validatePageConfig(basePath, data, fs);
+
+    const manifestPath = join(basePath, 'webapp/manifest.json');
+    const manifest = fs.readJSON(manifestPath) as Manifest;
+
+    const config = enhanceDataFn(data, manifest);
+
+    // enhance manifest.json
+    extendJSON(fs, {
+        filepath: manifestPath,
+        content: render(fs.read(getTemplatePath(templatePath)), config, {}),
+        replacer: getManifestJsonExtensionHelper(config),
+        tabInfo: data.tabInfo
+    });
 
     return fs;
 }
