+++ packages/fe-fpm-writer/src/page/list.ts
@@ -1,13 +1,7 @@
-import { join } from 'path';
-import { create as createStorage } from 'mem-fs';
 import type { Editor } from 'mem-fs-editor';
-import { create } from 'mem-fs-editor';
-import { render } from 'ejs';
-import { getFclConfig, getManifestJsonExtensionHelper, validatePageConfig } from './common';
+import { getFclConfig, extendPageJSON } from './common';
 import type { Manifest } from '../common/types';
-import { extendJSON } from '../common/file';
 import type { ListReport, InternalListReport } from './types';
-import { getTemplatePath } from '../templates';
 
 /**
  * Enhances the provided list report configuration with default data.
@@ -47,23 +41,5 @@ function enhanceData(data: ListReport, manifest: Manifest): InternalListReport {
  * @returns the updated memfs editor instance
  */
 export function generate(basePath: string, data: ListReport, fs?: Editor): Editor {
-    if (!fs) {
-        fs = create(createStorage());
-    }
-    validatePageConfig(basePath, data, fs);
-
-    const manifestPath = join(basePath, 'webapp/manifest.json');
-    const manifest = fs.readJSON(manifestPath) as Manifest;
-
-    const config = enhanceData(data, manifest);
-
-    // enhance manifest.json
-    extendJSON(fs, {
-        filepath: manifestPath,
-        content: render(fs.read(getTemplatePath('page/list/manifest.json')), config, {}),
-        replacer: getManifestJsonExtensionHelper(config),
-        tabInfo: data.tabInfo
-    });
-
-    return fs;
+    return extendPageJSON(basePath, data, enhanceData, 'page/list/manifest.json', fs);
 }
